+++ hw3/Watch.java
@@ -11,7 +11,7 @@ public class Watch extends eDevice{
      * @param quantity
      * @return void
      */
-    public Watch(String category, String name, int price, int quantity){
+    public Watch(String category, String name, double price, int quantity){
         super(category, name, price, quantity);
     }
 }
